SLM-21 Split integration tests to run separately from unit tests
