+++ build.gradle.kts
@@ -1,7 +1,12 @@
 plugins {
   id("uk.gov.justice.hmpps.gradle-spring-boot") version "3.3.13"
+  id("org.unbroken-dome.test-sets") version "4.0.0"
   kotlin("plugin.spring") version "1.6.0"
   kotlin("plugin.jpa") version "1.6.0"
+}
+
+testSets {
+  "testIntegration"()
 }
 
 configurations {
